[refactor/#50] import문 수정
