+++ src/main/java/com/clokey/server/domain/MemberTerm/application/MemberTermRepositoryService.java
@@ -1,6 +1,6 @@
 package com.clokey.server.domain.MemberTerm.application;
 
-import com.clokey.server.domain.model.mapping.MemberTerm;
+import com.clokey.server.domain.model.entity.mapping.MemberTerm;
 
 public interface MemberTermRepositoryService {
     void saveMemberTerm(MemberTerm memberTerm); // MemberTerm 저장
