+++ src/main/java/com/clokey/server/domain/category/converter/CategoryConverter.java
@@ -1,7 +1,7 @@
 package com.clokey.server.domain.category.converter;
 
 import com.clokey.server.domain.category.dto.CategoryResponseDTO;
-import com.clokey.server.domain.model.Category;
+import com.clokey.server.domain.model.entity.Category;
 import org.springframework.stereotype.Component;
 
 @Component
