+++ src/main/java/com/clokey/server/domain/cloth/application/ClothRepositoryService.java
@@ -1,6 +1,6 @@
 package com.clokey.server.domain.cloth.application;
 
-import com.clokey.server.domain.model.enums.Visibility;
+import com.clokey.server.domain.model.entity.enums.Visibility;
 
 public interface ClothRepositoryService {
