+++ src/main/java/com/clokey/server/domain/folder/application/FolderService.java
@@ -1,7 +1,7 @@
 package com.clokey.server.domain.folder.application;
 
 import com.clokey.server.domain.folder.dto.FolderRequest;
-import com.clokey.server.domain.model.Folder;
+import com.clokey.server.domain.model.entity.Folder;
 
 public interface FolderService {
     Folder createFolder(Long memberId, FolderRequest.FolderCreateRequest request);
