+++ src/main/java/com/clokey/server/domain/folder/application/FolderServiceImpl.java
@@ -4,8 +4,8 @@ import com.clokey.server.domain.folder.converter.FolderConverter;
 import com.clokey.server.domain.folder.dto.FolderRequest;
 import com.clokey.server.domain.folder.exception.FolderDeleteException;
 import com.clokey.server.domain.member.application.MemberRepositoryService;
-import com.clokey.server.domain.model.Folder;
-import com.clokey.server.domain.model.Member;
+import com.clokey.server.domain.model.entity.Folder;
+import com.clokey.server.domain.model.entity.Member;
 import com.clokey.server.global.common.response.BaseResponse;
 import com.clokey.server.global.error.code.status.ErrorStatus;
 import lombok.RequiredArgsConstructor;
