+++ src/main/java/com/clokey/server/domain/history/api/HistoryRestController.java
@@ -16,8 +16,8 @@ import com.clokey.server.domain.history.exception.validator.CommentValidator;
 import com.clokey.server.domain.history.exception.validator.HistoryAccessibleValidator;
 import com.clokey.server.domain.history.exception.validator.HistoryLikedValidator;
 import com.clokey.server.domain.member.exception.annotation.MemberExist;
-import com.clokey.server.domain.model.Comment;
-import com.clokey.server.domain.model.History;
+import com.clokey.server.domain.model.entity.Comment;
+import com.clokey.server.domain.model.entity.History;
 import com.clokey.server.global.common.response.BaseResponse;
 import com.clokey.server.global.error.code.status.SuccessStatus;
 import io.swagger.v3.oas.annotations.Operation;
