+++ src/main/java/com/clokey/server/domain/history/application/HistoryRepositoryService.java
@@ -1,7 +1,7 @@
 package com.clokey.server.domain.history.application;
 
-import com.clokey.server.domain.model.Comment;
-import com.clokey.server.domain.model.History;
+import com.clokey.server.domain.model.entity.Comment;
+import com.clokey.server.domain.model.entity.History;
 
 import java.util.List;
 import java.util.Optional;
