+++ src/main/java/com/clokey/server/domain/member/application/MemberRepositoryService.java
@@ -1,6 +1,6 @@
 package com.clokey.server.domain.member.application;
 
-import com.clokey.server.domain.model.Member;
+import com.clokey.server.domain.model.entity.Member;
 
 
 import java.util.Optional;
