+++ src/main/java/com/clokey/server/domain/member/application/MemberRepositoryServiceImpl.java
@@ -1,8 +1,7 @@
 package com.clokey.server.domain.member.application;
 
-import com.clokey.server.domain.member.dao.MemberRepository;
-import com.clokey.server.domain.model.Member;
-import com.clokey.server.domain.term.exception.TermException;
+import com.clokey.server.domain.model.repository.MemberRepository;
+import com.clokey.server.domain.model.entity.Member;
 import com.clokey.server.domain.member.exception.MemberException;
 import com.clokey.server.global.error.code.status.ErrorStatus;
 import jakarta.persistence.EntityManager;
