+++ src/main/java/com/clokey/server/domain/term/application/TermCommandServiceImpl.java
@@ -2,9 +2,9 @@ package com.clokey.server.domain.term.application;
 
 import com.clokey.server.domain.MemberTerm.application.MemberTermRepositoryService;
 import com.clokey.server.domain.member.application.MemberRepositoryService;
-import com.clokey.server.domain.model.Member;
-import com.clokey.server.domain.model.Term;
-import com.clokey.server.domain.model.mapping.MemberTerm;
+import com.clokey.server.domain.model.entity.Member;
+import com.clokey.server.domain.model.entity.Term;
+import com.clokey.server.domain.model.entity.mapping.MemberTerm;
 import com.clokey.server.domain.term.dto.TermRequestDTO;
 import com.clokey.server.domain.term.dto.TermResponseDTO;
 import com.clokey.server.domain.term.exception.TermException;
