+++ src/main/java/com/clokey/server/domain/term/application/TermRepositoryService.java
@@ -1,6 +1,6 @@
 package com.clokey.server.domain.term.application;
 
-import com.clokey.server.domain.model.Term;
+import com.clokey.server.domain.model.entity.Term;
 
 public interface TermRepositoryService {
